@@ -6,9 +6,9 @@ authors = [
     {name = "null2264", email = "palembani@gmail.com"}
 ]
 license = {text = "MPL-2.0"}
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    "fastapi==0.92.0",
+    "fastapi==0.100.0",
     "discord.py>=2.2",
     "pyzmq==25.0.0",
     "websockets>=10.4",
@@ -18,6 +18,7 @@ dependencies = [
     "oauthlib>=3.2.2",
     "beautifulsoup4>=4.12.2",
     "Pillow>=10.0.0",
+    "pydantic>=1.10,<=2.0",
 ]
 
 [tool.pdm.scripts]
